feat: show prompt for add integration and add provider commands
