+++ packages/cli/src/commands/add/index.ts
@@ -1,23 +1,34 @@
 import { Argument, Command } from "commander";
 
-import { Provider, createProviders, providerChoices } from "./create-provider";
-import { createResources } from "./create-resource";
 import { Integration, availableIntegrations } from "./add-integration";
+import {
+    Provider,
+    availableProviders,
+    createProviders,
+} from "./create-provider";
+import { createResources } from "./create-resource";
 import { integrateAntDesign } from "./integrations/ant-design";
 import { integrateReactRouter } from "./integrations/react-router";
-import { AddCommandPrompt, addCommandPrompt } from "./prompt";
+import {
+    AddCommandAnswers,
+    addCommandPrompt,
+} from "./prompts/add-command.prompt";
 
 const load = (program: Command) => {
     return program
         .command("add")
         .allowExcessArguments(false)
-        .addArgument(
-            new Argument("[providers...]", "Create provider(s)").choices(
-                providerChoices,
-            ),
-        )
-        .option("-p, --path [path]", "Path to generate providers")
         .action(addCommandAction)
+        .addCommand(
+            new Command("provider")
+                .addArgument(
+                    new Argument("[providers...]", "Create provider(s)")
+                        .choices(availableProviders)
+                        .default([]),
+                )
+                .option("-p, --path [path]", "Path to generate providers")
+                .action(createProviderAction),
+        )
         .addCommand(
             new Command("resource")
                 .addArgument(
@@ -34,7 +45,7 @@ const load = (program: Command) => {
         .addCommand(
             new Command("integration")
                 .addArgument(
-                    new Argument("<name>", "Name of the integration").choices(
+                    new Argument("[name]", "Name of the integration").choices(
                         availableIntegrations,
                     ),
                 )
@@ -55,27 +66,16 @@ const addCommandAction = async (
     await handlePromptResult(promptResult);
 };
 
-const handlePromptResult = async (promptResult: AddCommandPrompt) => {
-    if (promptResult.component === "provider") {
-        return createProviders(
-            promptResult.providers,
-            promptResult.providersPath,
-        );
-    }
-
-    if (promptResult.component === "integration") {
-        return await addIntegrationAction(promptResult.integration);
-    }
-
-    if (promptResult.component === "resource") {
-        return await createResources({ path: promptResult.resourcePath }, []);
-    }
-};
-
 const createProviderAction = async (
     providers: Provider[],
     options: { path?: string },
 ) => {
+    if (!providers.length) {
+        const promptResult = await addCommandPrompt({ component: "provider" });
+
+        return await handlePromptResult(promptResult);
+    }
+
     createProviders(providers, options.path);
 };
 
@@ -93,13 +93,38 @@ const createResourceAction = async (
     );
 };
 
-const addIntegrationAction = async (name: Integration) => {
+const addIntegrationAction = async (name?: Integration) => {
+    if (!name) {
+        const promptResult = await addCommandPrompt({
+            component: "integration",
+        });
+
+        await handlePromptResult(promptResult);
+    }
+
     if (name === "ant-design") {
         await integrateAntDesign();
     }
 
     if (name === "react-router") {
         await integrateReactRouter();
+    }
+};
+
+const handlePromptResult = async (promptResult: AddCommandAnswers) => {
+    if (promptResult.component === "provider") {
+        return createProviders(
+            promptResult.providers,
+            promptResult.providersPath,
+        );
+    }
+
+    if (promptResult.component === "integration") {
+        return await addIntegrationAction(promptResult.integration);
+    }
+
+    if (promptResult.component === "resource") {
+        return await createResources({ path: promptResult.resourcePath }, []);
     }
 };
